requirements.txt
fastapi==0.133.1
joblib==1.5.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
numpy==2.4.2
pandas==3.0.1
plotly==6.5.2
pydantic==2.12.5
PyYAML==6.0.3
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
uvicorn==0.41.0